Moved logging to a separate module
