+++ infoset/agents/standard/sentry3.py
@@ -16,9 +16,9 @@ from collections import defaultdict
 # infoset libraries
 from infoset.agents import agent
 from infoset.utils import jm_configuration
-from infoset.utils import jm_general
 from infoset.snmp import snmp_manager
 from infoset.snmp import mib_sentry3
+from infoset.log import log
 
 logging.getLogger('requests').setLevel(logging.WARNING)
 logging.basicConfig(level=logging.DEBUG)
@@ -96,7 +96,7 @@ class PollingAgent(object):
                 log_message = (
                     'No valid SNMP configuration found '
                     'for host "%s" ') % (hostname)
-                jm_general.log2quiet(1006, log_message)
+                log.log2quiet(1006, log_message)
                 continue
 
             # Create Query make sure MIB is supported
@@ -106,7 +106,7 @@ class PollingAgent(object):
                 log_message = (
                     'The Sentry3 MIB is not supported by host  "%s"'
                     '') % (hostname)
-                jm_general.log2quiet(1001, log_message)
+                log.log2quiet(1001, log_message)
                 continue
 
             # Get the UID for the agent after all preliminary checks are OK
