+++ infoset/interfaces/jm_cli.py
@@ -242,6 +242,7 @@ def _cli_pagemaker(subparsers, width=80):
             'Verbose Output.', width=width)
     )
 
+
 def _cli_monitor(subparsers, width=80):
     """Process "pagemaker" CLI commands.
