+++ infoset/log/__init__.py
@@ -1,0 +1,1 @@
+"""Infoset log package."""
