+++ infoset/utils/daemon.py
@@ -1,10 +1,10 @@
 """Generic linux daemon base class for python 3.x."""
 
+import atexit
+import signal
 import sys
 import os
 import time
-import atexit
-import signal
 
 
 class Daemon:
